// UIAngular/ClientApp/src/app/core/backend/models/message.ts
/* tslint:disable */
/* eslint-disable */
export interface Message {
  chatId?: number;
  content?: string;
  createdate?: string;
  id?: number;
  read?: boolean;
  reciverId?: string;
  reciverImg?: string;
  reciverName?: string;
  senderId?: string;
  senderImg?: string;
  senderName?: string;
}